Update AbortController to use the platform version
